Changed the way first frame death is prevented
Formerly, collision detection wasn't performed on the first frame. Now,
each enemy's hit box is updated when it is constructed, so that each enemy
doesn't start at the bottom left. I made this change because I don't want
newly spawned enemies to be able to kill the player by somehow spawning at
the bottom left of the screen.