Upgraded minimum API to 16 from 15
I'm not too highly concerned about the minimum API anyway, but I do like
to choose an API such that the number of users who could use the app is
maximized. However, I want to be able to use MediaPlayer.selectTrack().
For reference, my smartphone, on which I test, supports up to API 19.