The cheap fix worked
Thus, the background music switches from King Galaxian to Saber Wing
(these names are the tracks' real names) after the first level. The cheap
fix seems okay; I'm betting it's inefficient, but that should be fine,
since the background music is only switched in between (and thus not
during) levels.