Started revising the class hierarchy
For this commit, I created a class called WaypointEnemy that IS-A
EnemyEntity. To test this new class, I changed class Hunter so that it
IS-A WaypointEnemy. My plan is for the proposed enemies that need
waypoints (e.g. the ambusher) to extend WaypointEnemy and for the ones
that don't to inherit directly from EnemyEntity. I'm not sure whether to
make WaypointEnemy abstract, though. Part of me thinks it makes sense, but
part of me thinks it's an odd move, even though EnemyEntity is abstract. I
don't think this decision will be consequential, though.