Implemented all but Ambusher's behavior and sprite
I can't decide on a good way to implement the behavior. I erroneously
thought that it would be simple. Thus, since I really want to focus more
on my Supply and Demand Graph project and am thus trying to finish this
one quickly, I'm cancelling the Ambusher enemy (at least until I come back
to working on this app, if that ever happens).